Fix for cycle background color not working in fullscreen, generic shortcut fix
The new feature cycle background colors not being invoked in fullscreen because
the actions were not being added to the ImageWidget, since there's no menubar to
route them in fullscreen.

This commit removes the duplicated code adding actions to the ImageWidget and
adds a function createGlobalAction that takes care of that, so in the future this
bug won't reappear.

Files:

imageviewer.py:
- Created createGlobalAction, called for creating actions that need to work in
  fullscreen.
- Removed old code that replicated addAction for every fullscreen action

Tests:
- Loaded image files, verified keyboard shortcuts work in fullscreen and windowed