Removed thumbnail timer, added background decoding, moved Python threads to QThread
Before this commit, thumbnails were updated in a periodic UI timer callback that
did nothing if thumbnails were not dirty, scheduled the thumnbail file prefetch,
or blocked the UI decoding the thumbnail and updating the thumbnail widget. To
prevent blocking the UI, this required per updated thumbnail calls to
processEvents.

In addition, decoding of main images can take considerable time for 4K images or
svg, blocking the UI.

This commit adds decoding on a QThread for both thumnbail and main image. The
decoder thread can't just be a Python thread since it needs to access the
QImageReader object and generate QPixmap and emit a signal.

Python threads cannot emit signals, which are necessary for real asynchronous
handling without periodic timers (event driven vs. polling), so moved all Python
threads to QThread with emits and verified they don't hold the GIL when calling
internal functions that should release the GIL (QImageReader decoding, file
reading).

Threading the QImageReader uncovered a deadlock in the pyqt implementation of
QImageReader where Qt grabs a global lock to perform image plugin reader
initialization on the first method invoked on a given QImageReader instance.
This can cause deadlocks when that global lock is owned by a thread that is
waiting on the GIL and the thread owning the GIL is waiting on that global lock
(deadly priority inversion). The workaround is to serialize the call to the
first method of the QImageReader using a Python lock, which is done in
qThreadSafeImageReader.

Files:

imageviewer.py:
- Worked around deadlock due to Qt delayed initialization of QImageReader in
  qSafeImageReader
- Changed animation_interval_ms to be the time in between image decodings rather
  than a delay after decoding the image. If a image is not decoded in the given
  interval, it will be delayed another full interval (or more).
- Added image and thumbnail placeholders (error, loading, decoding)
- Changed prefetch_files to a FileFetcher QThread so it can emit signals that
  are handled on the UI thread.
- Added PixmapReader QThread to decode pixmaps from a queue via a QImageReader
- Added IMAGE_STATE states to simplify image and thumbnail loading/decoding
  state tracking
- Removed unnecessary prefetch_response queue since the UI thread is now
  notified via Qt signals
- Added signal handlers receive_file and receive_pixmap to receive the file data
  and the decoded pixmap respectively from the QThreads
- Image state is attached manually to ImageWidget objects (main image or
  thumbnails)
- Added image placeholders to use in the different states (queued for loading,
  loading, queued for decoding, decoding, error).
- Removed thumbnail timer used to refresh the UI when loaded, since now
  thumbnails are loaded and decoded in the background and the UI thread notified
  via a Qt signal.
- Stopped animation and prefetching when opening a file dialog box so the dialog
  box listdir is not blocked by image loads.
- Changed getDataFromCache to just put requests in the prefetch_request queue
- Changed updateThumbnails to progress the thumnbail IMAGE_STATE as needed
- Changed loadImage to either cause the file to be fetched or the pixmap to be
  decoded.
- Added updateImageData to queue into the decode thread using the file data
  fetched by the fetcher background thread. This will use a new reader or
  recycle an existing reader if this is an animated image.
- Moved all status line update to updateStatus

README.md
- Updated features

Tests:
- Loaded images, verified placeholders show in thumbnails and images
- Loaded GIF images, verified fastly switching across images doesn't hang,
  verified falling behind interval doesn't cause error
- Verified UI is not blocked when navigating thumnails, main image.