Quick fix for typo when accessing queue items causing exception, changed default logging to WARNING
The previous commit unified all the queue items to use an (id, payload) scheme
where payload was a tuple. This allowed several places that only cared about the
id to be independent of the number of elements in the payload.

Unfortunately there was a place in getDataFromCache that missed that
replacement. This commit fixes it.

Also changed default debugging to WARNING instead of INFO. This helps
updateThumbnails not to noticeably block the UI due to all the logging performed
in that function and related decoding threads (which, if they are able to sneak
in due to the GIL being released for eg thumbWidget rescaling, can block the
UI thread due to the GIL).

Files:

imageviewer.py:
- Fixed getDataFromCache to use the id, payload scheme when clearing the queue
- Changed default logging to WARNING

Tests:
- Displayed lists of local and remote images, verified no exceptions due to
  fetching the wrong number of items from queues.